+++ cider-mode.el
@@ -213,12 +213,12 @@ the related commands `cider-repl-clear-buffer' and
 (defconst cider-mode-eval-menu
   '("CIDER Eval" :visible (cider-connected-p)
     ["Eval top-level sexp" cider-eval-defun-at-point]
-    ["Eval top-level sexp to point" cider-eval-defun-to-point]
+    ["Eval top-level sexp to point" cider-eval-defun-up-to-point]
     ["Eval top-level sexp to comment" cider-eval-defun-to-comment]
     ["Eval top-level sexp and pretty-print to comment" cider-pprint-eval-defun-to-comment]
     "--"
     ["Eval current sexp" cider-eval-sexp-at-point]
-    ["Eval current sexp to point" cider-eval-sexp-to-point]
+    ["Eval current sexp to point" cider-eval-sexp-up-to-point]
     ["Eval current sexp in context" cider-eval-sexp-at-point-in-context]
     "--"
     ["Eval last sexp" cider-eval-last-sexp]